Add Widget in main area, render a cube, reduce the num of libs
Add a splitter in the main area of the window,and put a tabView in the
left of splitter and a QVTKWidget in the right half.
Add a temp function 'displayCube' in class MainWindow as a demo.
delete the libs of vtk and Qt that are not nesessary to reduce
the memory use.

GVSClient-3