Finish the feature of adding legend item
Fix the recursively invoke of displaying the Addition Dialog.
Fix the refreshing of editor list after adding an item.
Fix the issue of adding item when clicking cancle button.
Fix the issue of display the wrong color in the legend when
the value of r or g or b is 255, by change the mod base to 256.

GVSClient-13