Not finished, add the feature of downloading project
Issue of unable to get the list of projects in the root directory.
when execute 'list()' in root path, the signal listInfo() does not
call the slot 'addToList()'. Do not know the reason.
Maybe it is that Qt is not fully supporting for win7, but other client
of FTP is ok when reading 127.0.0.1:21, and could see the full list

of the current directory.

GVSClient-15