Add the feature of editting legend, some optmization of lengend manager
Add palette when add and edit legend.
Change the format of legend in gvp file.
And change the way it parsing out legend record.
Delete the id of legend record, and treat the name as the unique identifier.
Change the default path when open a project.

Add the feature of editting legend existed.

GVSClient-28